requirements.txt
# MERLIN Python dependencies (core + optional subsystems)
# Canonical source remains pyproject.toml; this file is manually synchronized.

pydantic>=2.7,<3.0
PyYAML>=6.0,<7.0
python-dotenv>=1.0,<2.0
prompt-toolkit>=3.0,<4.0
requests>=2.31,<3.0
ruamel.yaml>=0.18,<0.19
psutil>=5.9,<8.0

# Testing
pytest>=8.2,<9.0
numpy>=1.26,<3.0

# API server
fastapi>=0.111,<1.0
uvicorn>=0.30,<1.0

# Voice I/O
sounddevice>=0.4,<1.0
webrtcvad>=2.0.10,<3.0
pyttsx3>=2.90,<3.0
faster-whisper>=1.0,<2.0

# UI + browser integrations
PySide6>=6.7,<7.0
browser-use>=0.1.48,<1.0
langchain-google-genai>=2.0,<3.0

# Windows-only integrations
pywin32>=306; platform_system == 'Windows'
comtypes>=1.2,<2.0; platform_system == 'Windows'
pycaw>=20240210; platform_system == 'Windows'
pygetwindow>=0.0.9; platform_system == 'Windows'
winsdk>=1.0; platform_system == 'Windows'

# Telegram bot
python-telegram-bot>=21.0,<22.0
